requirements.txt
fastapi==0.97.0
pydantic==1.10.9
requests==2.31.0
aiohttp==3.9.3
alembic==1.13.1
python-dotenv==1.0.0
uvicorn==0.22.0
psycopg2==2.9.9
PyJWT==2.4.0
passlib==1.7.4
bcrypt==3.2.0